pyproject.toml
[project]
name = "ai-web-search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "lxml-html-clean>=0.4.2",
    "newspaper3k>=0.2.8",
    "nltk>=3.9.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.1",
    "torch>=2.7.1",
    "transformers>=4.54.1",
    "watchdog>=6.0.0",
]
